pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clotho-engine"
version = "0.1.0"
description = "HFT-Grade Deterministic Simulation & Fuzzing Engine"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Enting Zhang"}
]
keywords = [
    "simulation",
    "testing",
    "distributed-systems",
    "fuzzing",
    "deterministic",
    "chaos-engineering"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Distributed Computing",
]
requires-python = ">=3.10"
dependencies = [
    "lark>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/e35zhang/Clotho-Simulation-Engine"
Repository = "https://github.com/e35zhang/Clotho-Simulation-Engine"
Issues = "https://github.com/e35zhang/Clotho-Simulation-Engine/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["core"]
omit = ["tests/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
]
